- Util 클래스들에 메소드 추가 - GalleryFragment 어디에서든 쓰일 수 있게 리팩토링

+++ app/src/main/java/com/angel/black/baskettogether/recruit/RecruitPostListActivity.java
@@ -111,7 +111,7 @@ public class RecruitPostListActivity extends BaseListActivity implements View.On
         fab.setRippleColor(getResources().getColor(R.color.colorPrimaryDark));
         fab.setOnClickListener(this);
         fab.setTag(TAG_FAB);
-        int size = (int) ScreenUtil.convertDpToPixel(this, 48);
+        int size = ScreenUtil.convertDpToPixel(this, 48);
         FrameLayout.LayoutParams params = new FrameLayout.LayoutParams(size, size, Gravity.RIGHT|Gravity.BOTTOM);
 
         params.bottomMargin = size/3;
